ref: create useCookieListing hook to handle table related props
